вынес переменные и словари в const
