+++ src/const.js
@@ -1,6 +1,8 @@
 const DEFAULT_EVENT_TYPE = 'flight';
 const EVENT_TYPES = ['taxi', 'bus', 'train', 'ship', 'drive', 'flight', 'check-in', 'sightseeing', 'restaurant'];
 const DESTINATIONS_ITEMS_COUNT = 3;
+const AUTHORIZATION = 'Basic umB8tRoVHKu5p3GdQ';
+const END_POINT = 'https://21.objects.pages.academy/big-trip';
 
 const POINT_EMPTY = {
   basePrice: 0,
@@ -71,8 +73,15 @@ const SourceUrl = {
   DESTINATIONS: 'destinations',
 };
 
+const TimeLimit = {
+  LOWER_LIMIT: 350,
+  UPPER_LIMIT: 1000,
+};
+
 export {
   DESTINATIONS_ITEMS_COUNT,
+  AUTHORIZATION,
+  END_POINT,
   EVENT_TYPES,
   POINT_EMPTY,
   FilterType,
@@ -85,4 +94,5 @@ export {
   EditType,
   Method,
   SourceUrl,
+  TimeLimit,
 };
